Rename Ctype -> Type.
C in Ctype stands for the C language because it represents a type in C.
That is however redundant because this is a C compiler after all and
all type information that the compiler needs to handle is about C.
We should drop C from Ctype to make it more readable.

The existing identifier "type" would conflict with the renamed type,
so it's renamed "kind".