Make fulltest faster.
Make fulltest compiles 8cc three times. The first executable is compiled
using the system's C compiler. The second and the third are compiled using
8cc itself. And then we verify the second and the third executables are
the same (to verify that the 8cc outputs reached fixed point).

If the third and the second executables are the same, their test results
should be the same. So we don't need to run the unit tests for the third-gen
executable.