Improve the regsave code for the K&R-style function.
If a function takes variable arguments, the compiler emits the
code to save registers being used to pass parameters to the stack,
so that the subsequent va_start and va_arg can access the variable
arguments.

8cc did that even for a function declared as foo() (as opposed to
foo(void)). That's not necessary. You cannot call va_start in such
function because your function has no declared parameter (va_start
needs the argument before the ellipsis.)

This patch makes 8cc to not emit register save code for the K&R-style
function. This does not affect the correctness, but will make the
executable smaller.