Cleanup: rename "tostring" functions.
Short names are generally good but they are too short.
Reanme a2s -> node2s, c2s -> ty2s, and t2s -> tok2s.

I tried to use C11 _Generic to define a generic "tostring" macro
like this.

#define S(x) _Generic((x), Node*:node2s, Type*:ty2s, Token*:tok2s)(x)

However this didn't compile with GCC 4.8.2 (Clang 3.4 did).
It may also be confusing to those who are not familiar with the new
C11 feature. So I decided to not use that.