Fix continuation line bug.
We allowed space characters between backslash and newline because
both GCC and Clang allows them with warning. But there was a bug
in 8cc that replaces sequnce of space characters with one space.

The C standard doesn't allow such spaces, so I decided to remove
the feature instead of fixing its behavior. There's no clean way
to fix the issue because it would require reading ahead unbounded
number of characters.