Remove "autopop" field from File struct.
Sometimes we want to read from streams unrelated to the main input file.
For example, we want to use the parser to read a token from a string
after concatenating token strings with ##.

We had "autopop" field for each File struct to prevent temporary inputs
from being mixed with the main input stream. I don't like the way how
it works. So I rewrite it.