Fix bug in __TIMESTAMP__.
Previously, we call fstat on current_file() for __TIMESTAMP__ macro.
That may produce inaccurate results because the current file might be
different from a file that a token was created from.

In this patch, we call fstat on each FILE object and store the
result to a File object. When you need a file's modified time,
you can always get that by accessing File's mtime field.

We cannot call fstat when we need a timestamp because a file
that produced a token might have already been closed.