Parser: Print accurate error locations.
Previously, the parser prints out the current file location when
printing out an error message. The current file location is always
a bit ahead (or something far ahead) of a token that cuased an error.
So, the source location was less accurate.

In this patch, I added a new error function (actually a macro), errort().
errort() takes a token which caused an error. Because tokens have
original source locations, we can print out exact locations of errors.